+++ docs/changelog.md
@@ -8,6 +8,10 @@ icon: history
 
 # Changelog
 
+## 2026-07-18
+
+- Removed starter sample API content from the main documentation source.
+
 ## 2026-07-17
 
 - Created the initial Dokumentasi documentation foundation.